program replacing specific words by another one
from a given text for test purpose, but needs to be linked to the whole program. 

Only find a word and replace it for the moment. Need to add conditions about when to replace it or not, images, change the Content-Lenght Header. 
Need to add a "memory" in case of the word being separated in two segments.